archive/codeswarm_legacy/requirements.txt
google-genai
google-adk[extensions]
a2a-sdk
python-dotenv
requests
beautifulsoup4